requirements.txt
google-generativeai
pillow
streamlit
streamlit-option-menu